Extract shared NDS bucket resolution into a reusable concern
ButtonComponent (#13393) and AlertComponent both resolved the render-time
bucket with an identical private nds_bucket? plus a before_render that set
@render_as_nds. Move that shared logic into NDSBucketResolvable and have both
components include it, so the resolution lives in one place.

Each component keeps its own nds_delegate and supplies an nds_variant_class
hook naming the variant to exclude from the flip, preserving existing behavior
exactly.

changelog: Internal, Components, Extract shared NDS bucket resolution into a reusable concern
